refactor: Card component index에서 분리
